rag-react-appNew/backend/app/config.py
import os
from dotenv import load_dotenv

load_dotenv()

PDF_DIR = "data/pdfs"
CHROMA_DIR = "data/chroma"

OLLAMA_CHAT_MODEL = os.getenv("OLLAMA_CHAT_MODEL", "llama3.1")
OLLAMA_EMBED_MODEL = os.getenv("OLLAMA_EMBED_MODEL", "nomic-embed-text")
OLLAMA_BASE_URL = os.getenv("OLLAMA_BASE_URL", "http://localhost:11434")