rag-react-appNew/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-ollama
langchain-chroma
chromadb
pymupdf
python-dotenv